:ambulance: Fix bug in reorder middleware, add go template middleware
